+++ CHANGELOG.md
@@ -7,3 +7,4 @@
 - basic math expression: `+`, `-`, `*`, `/`, `^`
 - each expression is separated by semicolons
 - printing of math expressions using the `print` keyword
+- printing of numbers interpreted as ascii characters using the `print_char` keyword